requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.17.1
kombu==5.6.2
packaging==26.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.3.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
